+++ server/internal/archive/dimag/control.go
@@ -3,58 +3,87 @@ package dimag
 import (
 	"context"
 	"encoding/xml"
-	"lath/xman/internal/archive/shared"
 	"lath/xman/internal/db"
 	"path/filepath"
 )
 
-const XmlHeader = "<?xml version='1.0' encoding='UTF-8'?>\n"
-const ControlFileName string = "controlfile.xml"
-const InformationObjectAbbreviation string = "O"
-const RepresentationAbbreviation string = "R"
-const FileAbbreviation string = "F"
+var XmlHeader = []byte("<?xml version='1.0' encoding='UTF-8'?>\n")
 
-func GenerateControlFile(
+type itemType string
+
+const (
+	itemTypeInformationObject itemType = "O"
+	itemTypeRepresentation    itemType = "R"
+	itemTypeFile              itemType = "F"
+)
+
+type dimagControl struct {
+	XMLName    xml.Name `xml:"verzeichnungseinheit"`
+	RootID     string   `xml:"rootid"`
+	IndexItems []indexItem
+}
+
+type indexItem struct {
+	XMLName    xml.Name `xml:"verz-obj"`
+	IndexID    string   `xml:"aid"`
+	Lifetime   string   `xml:"entstehungs-zeitraum"`
+	FilePath   string   `xml:"sftp-dateiname,omitempty"`
+	FileName   string   `xml:"dateiname"`
+	Title      string   `xml:"titel"`
+	ItemType   itemType `xml:"typ"`
+	IndexItems []indexItem
+}
+
+func generateControlFile(
 	message db.Message,
-	archivePackageData db.ArchivePackage,
+	archivePackage db.ArchivePackage,
 	importDir string,
-) string {
-	primaryDocuments := archivePackageData.PrimaryDocuments
-	fileIndexItems := []IndexItem{}
-	for _, primaryDocument := range primaryDocuments {
-		fileIndexItems = append(
-			fileIndexItems,
-			GetIndexItemForPrimaryDocument(primaryDocument.PrimaryDocument, importDir),
-		)
+) []byte {
+	primaryDocuments := archivePackage.PrimaryDocuments
+	fileIndexItems := []indexItem{}
+	for _, d := range primaryDocuments {
+		item := indexItem{
+			ItemType: itemTypeFile,
+			Title:    fileName(d.PrimaryDocument),
+			FileName: fileName(d.PrimaryDocument),
+			FilePath: filepath.Join(importDir, d.Filename),
+		}
+		fileIndexItems = append(fileIndexItems, item)
 	}
-	xmlIndexItem := IndexItem{
-		IndexID:  "",
-		ItemType: FileAbbreviation,
+	messageIndexItem := indexItem{
+		ItemType: itemTypeFile,
 		Title:    filepath.Base(message.MessagePath),
-		FilePath: getRemoteXmlPath(message, importDir),
+		FilePath: filepath.Join(importDir, filepath.Base(message.MessagePath)),
 	}
-	protocolIndexItem := GetIndexItemForProtocol(importDir)
-	fileIndexItems = append(fileIndexItems, xmlIndexItem, protocolIndexItem)
-	repIndexItem := IndexItem{
-		IndexID:    "",
-		ItemType:   RepresentationAbbreviation,
-		Title:      archivePackageData.REPTitle,
+	fileIndexItems = append(fileIndexItems, messageIndexItem)
+	// protocolIndexItem := indexItem{
+	// 	ItemType: itemTypeFile,
+	// 	Title:    shared.ProtocolFilename,
+	// 	FileName: shared.ProtocolFilename,
+	// 	FilePath: filepath.Join(importDir, shared.ProtocolFilename),
+	// }
+	// fileIndexItems = append(fileIndexItems, protocolIndexItem)
+	repIndexItem := indexItem{
+		ItemType:   itemTypeRepresentation,
+		Title:      archivePackage.REPTitle,
 		IndexItems: fileIndexItems,
 	}
-	repItems := []IndexItem{repIndexItem}
-	ioIndexItem := IndexItem{
-		IndexID:    "",
-		Lifetime:   getCombinedLifetime(archivePackageData.IOLifetime),
-		ItemType:   InformationObjectAbbreviation,
-		Title:      archivePackageData.IOTitle,
+	repItems := []indexItem{repIndexItem}
+	ioIndexItem := indexItem{
+		Lifetime:   combinedLifetime(archivePackage.IOLifetime),
+		ItemType:   itemTypeInformationObject,
+		Title:      archivePackage.IOTitle,
 		IndexItems: repItems,
 	}
-	ioItems := []IndexItem{ioIndexItem}
-	archiveCollection, ok := db.FindArchiveCollection(context.Background(), archivePackageData.CollectionID)
+	ioItems := []indexItem{ioIndexItem}
+	archiveCollection, ok := db.FindArchiveCollection(
+		context.Background(),
+		archivePackage.CollectionID,
+	)
 	if !ok {
-		panic("failed to find archive collection: " + archivePackageData.CollectionID.Hex())
+		panic("failed to find archive collection: " + archivePackage.CollectionID.Hex())
 	}
-	dimagControl := DimagControl{
+	dimagControl := dimagControl{
 		RootID:     archiveCollection.DimagID,
 		IndexItems: ioItems,
 	}
@@ -62,51 +91,18 @@ func GenerateControlFile(
 	if err != nil {
 		panic(err)
 	}
-	controlFileString := XmlHeader + string(xmlBytes)
-	return controlFileString
+	return append(XmlHeader, xmlBytes...)
 }
 
-func GetIndexItemForPrimaryDocument(
-	primaryDocument db.PrimaryDocument,
-	importDir string,
-) IndexItem {
-	fileIndexItem := IndexItem{
-		IndexID:  "",
-		ItemType: FileAbbreviation,
-		Title:    getFileName(primaryDocument),
-		FileName: getFileName(primaryDocument),
-		FilePath: filepath.Join(importDir, primaryDocument.Filename),
-	}
-	return fileIndexItem
-}
-
-func GetIndexItemForProtocol(
-	importDir string,
-) IndexItem {
-	fileIndexItem := IndexItem{
-		IndexID:  "",
-		ItemType: FileAbbreviation,
-		Title:    shared.ProtocolFilename,
-		FileName: shared.ProtocolFilename,
-		FilePath: filepath.Join(importDir, shared.ProtocolFilename),
-	}
-	return fileIndexItem
-}
-
-func getFileName(primaryDocument db.PrimaryDocument) string {
+func fileName(primaryDocument db.PrimaryDocument) string {
 	if primaryDocument.FilenameOriginal == "" {
 		return primaryDocument.Filename
 	}
 	return primaryDocument.FilenameOriginal
 }
 
-func getRemoteXmlPath(message db.Message, importDir string) string {
-	filename := filepath.Base(message.MessagePath)
-	return filepath.Join(importDir, filename)
-}
-
-// getCombinedLifetime returns a string representation of lifetime start and end.
-func getCombinedLifetime(l *db.Lifetime) string {
+// combinedLifetime returns a string representation of lifetime start and end.
+func combinedLifetime(l *db.Lifetime) string {
 	if l != nil {
 		if l.Start != "" && l.End != "" {
 			return l.Start + " - " + l.End